refactor: CreateForm
