+++ pages/create/index.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Create() {
+  return (
+    <div>Create</div>
+  )
+}
