--- pages/home.js
@@ -1,15 +1,0 @@
-import React, { useEffect, useState } from "react";
-export default function Home() {  
-  return (
-    <div>
-      <a
-        onClick={() => {}}
-        target="_blank"
-        rel="noreferrer"
-        className="home__buttons__button button--primary"
-      >
-        Send
-      </a>
-    </div>
-  );
-}
